requirements.txt
flask
tensorflow
numpy
Pillow
scikit-learn
opencv-python-headless
